Add red-click-check to MouseUtil
Moved isolate_colors function from rcv to bcv
